% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCPhelpers.R
\name{sim_data_generator_1D_example1}
\alias{sim_data_generator_1D_example1}
\title{sim_data_generator_1D_example1 - generate 1D simulated data for example1}
\usage{
sim_data_generator_1D_example1(
  sim_name,
  n = 1000,
  n0 = 1000,
  m = 1000,
  alpha = 0.05
)
}
\description{
sim_data_generator_1D_example1 - generate 1D simulated data for example1
}
