% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCPhelpers.R
\name{sim_data_generator_1D_example2}
\alias{sim_data_generator_1D_example2}
\title{sim_data_generator_1D_example2 - generate 1D simulated data for example2}
\usage{
sim_data_generator_1D_example2(
  sim_name,
  n = 1000,
  n0 = 1000,
  m = 1000,
  alpha = 0.05
)
}
\description{
sim_data_generator_1D_example2 - generate 1D simulated data for example2
}
